dom/tests/mochitest/bugs/test_bug291653.html

<!DOCTYPE HTML>
<html>
<!--
https://bugzilla.mozilla.org/show_bug.cgi?id=291653
-->
<head>
  <title>Test for Bug 291653</title>
  <script type="application/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css"/>
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=291653">Mozilla Bug 291653</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script type="application/javascript">

/** Test for Bug 291653 **/
SimpleTest.waitForExplicitFinish();
SimpleTest.requestFlakyTimeout("untriaged");

var secondListenerDidRun = false;

var w;
function start() {
  if ("showModalDialog" in window) {
    w = window.open("file_bug291653.html", "foo", "width=300,height=300");
  } else {
    // window.showModalDialog doesn't exist in e10s mode, nothing to do in this test.
    ok(true, "nothing to do in e10s mode");
    SimpleTest.finish();
  }
}

function closeTest() {
  w.close();
  setTimeout("finish()", 500);
}

function finish() {
  ok(!secondListenerDidRun, "Shouldn't have run second listener!");
  SimpleTest.finish();
}

function end() {
  setTimeout("closeTest()", 500);
}

start();

</script>
</pre>
</body>
</html>
